+++ examples/lib/commons/square_component.dart
@@ -2,21 +2,20 @@ import 'dart:ui';
 
 import 'package:flame/components.dart';
 import 'package:flame/effects.dart';
-import 'package:flame/palette.dart';
+import 'package:flutter/material.dart';
 
-class SquareComponent extends PositionComponent with EffectsHelper {
-  Paint paint = BasicPalette.white.paint();
-
-  SquareComponent({int priority = 0, double size = 100.0})
-      : super(
-          size: Vector2.all(size),
+class SquareComponent extends RectangleComponent with EffectsHelper {
+  SquareComponent({
+    Vector2? position,
+    double size = 100.0,
+    Paint? paint,
+    int priority = 0,
+  }) : super(
+          Vector2.all(size),
+          position: position,
+          paint: paint,
           priority: priority,
-          anchor: Anchor.center,
-        );
-
-  @override
-  void render(Canvas c) {
-    super.render(c);
-    c.drawRect(size.toRect(), paint);
+        ) {
+    anchor = Anchor.center;
   }
 }