+++ examples/lib/stories/effects/combined_effect.dart
@@ -15,9 +15,7 @@ class CombinedEffectGame extends FlameGame with TapDetector {
   @override
   Future<void> onLoad() async {
     await super.onLoad();
-    greenSquare = SquareComponent()
-      ..paint = green
-      ..position.setValues(100, 100);
+    greenSquare = SquareComponent(position: Vector2.all(100))..paint = green;
 
     redSquare = SquareComponent()
       ..paint = red